休整了出动态和优惠券模块的其他模块

+++ api/src/main/java/com/wenjia/api/service/UserService.java
@@ -13,7 +13,7 @@ public interface UserService extends IService<User> {
     void register(UserDTO userDTO);
 
     //查询粉丝数量
-    Integer getFansNumberById(Integer id);
+    Integer getFansNumberById(Long id);
 
     //增加粉丝数
     void incrFansNumber(Long id);
